Dry run implemented

+++ pdf_sorter/file_manipulator.py
@@ -1,6 +1,7 @@
 import os
 import shutil
 import logging
+import settings
 
 logger = logging.getLogger(__name__)
 
@@ -17,10 +18,12 @@ def rename_file(path, old_name, config):
         logging.warning("Expected Token named " + exception + " but it was not found. Skipping rename.")
 
     logger.info("Renaming file to..." + new_name)
-    os.rename(path + "/" + old_name, path + "/" + new_name)
+    if settings.dry_run is False:
+        os.rename(path + "/" + old_name, path + "/" + new_name)
     return new_name
 
 
 def move_file(source_directory, file, target_directory):
     logger.info("Moving file to..." + target_directory)
-    shutil.move(source_directory + "/" + file, target_directory + "/" + file)
+    if settings.dry_run is False:
+        shutil.move(source_directory + "/" + file, target_directory + "/" + file)
